Modify sample.py to account for new folder scheme
Modified the original sample.py to get it running in the new folder
scheme for this project. The scheme is outlined in the README. Second
modification was to switch the code to using .tsv files over .pkl,
since those are the sources we have available.

APatK.py was untouched as it's useless on its own anyways. Without
the 'correct' comparison data it is purely an illustration of AP@k.

Closes task #14800544868586