Notebook for shinking the data sets
Full sets are too large, so to make training and testing more
manageable I've written a notebook to slice out about 1/4 of the
training data into a new training set. Another slice of the training
data is parsed to become the testing dataset.

The new sets will be uploaded to dropbox.